Refactor item selection code
