+++ lib/features/library/library_view.dart
@@ -2,6 +2,7 @@ import 'package:book_adapter/features/library/data/book_collection.dart';
 import 'package:book_adapter/features/library/data/item.dart';
 import 'package:book_adapter/features/library/library_view_controller.dart';
 import 'package:book_adapter/features/library/widgets/add_book_button.dart';
+import 'package:book_adapter/features/library/widgets/item_list_tile_widget.dart';
 import 'package:book_adapter/features/library/widgets/profile_button.dart';
 import 'package:book_adapter/localization/app.i18n.dart';
 import 'package:cached_network_image/cached_network_image.dart';
@@ -15,13 +16,13 @@ import 'package:sliver_tools/sliver_tools.dart';
 import 'package:sticky_headers/sticky_headers.dart';
 
 /// Displays a list of BookItems.
-class LibraryView extends ConsumerWidget {
+class LibraryView extends StatelessWidget {
   const LibraryView({ Key? key }) : super(key: key);
 
   static const routeName = '/';
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
+  Widget build(BuildContext context) {
     return const Scaffold(
       body: LibraryScrollView(),
     );
@@ -77,11 +78,14 @@ class LibraryScrollView extends HookConsumerWidget {
             ? isSelectingAppBar
             : notSelectingAppBar,
           duration: const Duration(microseconds: 15000),
-        ),// List of collections
+        ),
+        
+        // List of collections
         SliverImplicitlyAnimatedList<BookCollection>(
           items: data.collections ?? [],
           areItemsTheSame: (a, b) => a.id == b.id,
-          itemBuilder: (context, animation, collection, index) => collectionsBuilder(context, animation, collection, index, scrollController),
+          itemBuilder: (context, animation, collection, index) 
+              => collectionsBuilder(context, animation, collection, index, scrollController),
         ),
       ],
     );
@@ -89,6 +93,7 @@ class LibraryScrollView extends HookConsumerWidget {
 
   Widget collectionsBuilder(BuildContext context, Animation<double> animation, BookCollection collection, int index, ScrollController controller) {
     return StickyHeader(
+      key: ValueKey(collection.id),
       controller: controller,
       header: Container(
         height: 50.0,
@@ -99,18 +104,24 @@ class LibraryScrollView extends HookConsumerWidget {
           style: const TextStyle(color: Colors.white),
         ),
       ),
-      content: BookCollectionList(collection: collection),
+      content: BookCollectionList(
+        key: ValueKey(collection.id + 'BookCollectionList'),
+        collection: collection,
+      ),
     );
   }
 }
 
 class BookCollectionList extends HookConsumerWidget {
   const BookCollectionList({ Key? key, required this.collection }) : super(key: key);
+
   final BookCollection collection;
+
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     final LibraryViewData data = ref.watch(libraryViewController);
     final LibraryViewController viewController = ref.watch(libraryViewController.notifier);
+
     final items = data.books?.where((book) => book.collectionIds.contains(collection.id))
       .toList() ?? [];
     items.sort((a, b) => a.title.compareTo(b.title));
@@ -119,7 +130,6 @@ class BookCollectionList extends HookConsumerWidget {
       padding: const EdgeInsets.only(bottom: 16, top: 4, left: 8, right: 8),
       physics: const NeverScrollableScrollPhysics(),
       shrinkWrap: true,
-      key: ValueKey(collection.id),
       items: items,
       areItemsTheSame: (a, b) => a.id == b.id,
       itemBuilder: booksBuilder,
@@ -128,8 +138,6 @@ class BookCollectionList extends HookConsumerWidget {
   }
 
   Widget removeItemBuilder(BuildContext context, Animation<double> animation, Item oldItem, LibraryViewController viewController, LibraryViewData data) {
-    final imageUrl = oldItem.imageUrl;
-    final subtitle = oldItem.subtitle;
     final isSelected = data.selectedItemIds.contains(oldItem.id);
 
     if (isSelected) {
@@ -138,14 +146,12 @@ class BookCollectionList extends HookConsumerWidget {
     
     return FadeTransition(
       opacity: animation,
-      child: ListTile(
-        key: ValueKey(collection.id + oldItem.id),
-        title: Text(oldItem.title),
-        subtitle: subtitle != null ? Text(subtitle) : null,
-        leading: imageUrl != null 
-          ? ClipRRect(child: CachedNetworkImage(imageUrl: imageUrl, width: 40,), borderRadius: BorderRadius.circular(4),)
-          : null,
-      )
+      key: ValueKey(collection.id + oldItem.id + 'FadeTransition'),
+      child: ItemListTileWidget(
+        key: ValueKey(collection.id + oldItem.id + 'ItemListWidget'),
+        item: oldItem,
+        collection: collection,
+      ),
     );
   }
 
@@ -154,88 +160,12 @@ class BookCollectionList extends HookConsumerWidget {
       sizeFraction: 0.7,
       curve: Curves.easeInOut,
       animation: animation,
-      child: ItemListWidget(
+      child: ItemListTileWidget(
+        key: ValueKey(collection.id + item.id + 'ItemListWidget'),
         item: item,
+        collection: collection,
       ),
     );
   }
 }
 
-class ItemListWidget extends ConsumerWidget {
-  const ItemListWidget({
-    Key? key,
-    required this.item,
-  }) : super(key: key);
-
-  final Item item;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final LibraryViewData data = ref.watch(libraryViewController);
-    final LibraryViewController viewController = ref.watch(libraryViewController.notifier);
-
-    final imageUrl = item.imageUrl;
-    final subtitle = item.subtitle;
-    final isSelected = data.selectedItemIds.contains(item.id);
-
-    final Widget? image = imageUrl != null 
-      ? ClipRRect(
-          child: CachedNetworkImage(
-            imageUrl: imageUrl,
-            width: 40,
-          ),
-          borderRadius: BorderRadius.circular(4),
-        )
-      : null;
-
-
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 3),
-      child: Card(
-        margin: EdgeInsets.zero,
-        color: isSelected ? Colors.white30 : null,
-        elevation: 0,
-        child: Stack(
-          alignment: Alignment.center,
-          children: [
-            ListTile(
-              contentPadding: const EdgeInsets.symmetric(horizontal: 8),
-              shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
-              key: ValueKey(item.id),
-              title: Text(item.title),
-              subtitle: subtitle != null ? Text(subtitle) : null,
-              leading: image,
-              onLongPress: () => viewController.selectItem(item.id),
-              onTap: () {
-                if (isSelected) {
-                  return viewController.deselectItem(item.id);
-                }
-            
-            
-                if (data.isSelecting) {
-                  return viewController.selectItem(item.id);
-                }
-            
-                // Navigate to the reader page or series page depending on item type.
-                Navigator.restorablePushNamed(
-                  context,
-                  item.routeTo,
-                  arguments: item.toMapSerializable(),
-                );
-              },
-            ),
-            if (isSelected)
-              Positioned(
-                top: 8,
-                right: 8,
-                child: Icon(
-                  Icons.check_circle,
-                  color: Theme.of(context).canvasColor,
-                ),
-              ),
-          ],
-        ),
-      ),
-    );
-  }
-}
